+++ app/src/main/java/com/mobven/fitai/login/AuthFragment.kt
@@ -1,0 +1,32 @@
+package com.mobven.fitai.login
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.navigation.fragment.findNavController
+import com.mobven.fitai.R
+import com.mobven.fitai.databinding.FragmentAuthBinding
+
+class AuthFragment : Fragment() {
+
+    private lateinit var binding: FragmentAuthBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = FragmentAuthBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.registerButton.setOnClickListener {
+            findNavController().navigate(R.id.action_authFragment_to_signUpFragment)
+        }
+
+    }
+}
