--- app/src/main/java/com/mobven/fitai/login/sign_up/screens/SignUpFragment.kt
@@ -1,19 +1,0 @@
-package com.mobven.fitai.login.sign_up.screens
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.mobven.fitai.R
-
-class SignUpFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_sign_up, container, false)
-    }
-
-}
