+++ app/src/main/res/values/dimens.xml
@@ -27,4 +27,6 @@
     <dimen name="_20sp">20sp</dimen>
     <dimen name="_25dp">25dp</dimen>
     <dimen name="_26sp">26sp</dimen>
+    <dimen name="_5dp">5dp</dimen>
+    <dimen name="_3dp">3dp</dimen>
 </resources>
